requirements.txt
pandas
streamlit
plotly.express
scikit-learn